@@ -23,10 +23,17 @@
   </licenses>
 
   <properties>
-    <java.version>11</java.version>
-    <folio-spring-base-version>5.0.3</folio-spring-base-version>
+    <java.version>17</java.version>
+    <folio-spring-base-version>6.0.0-SNAPSHOT</folio-spring-base-version>
+    <versions-maven-plugin-version>2.13.0</versions-maven-plugin-version>
+    <maven-enforcer-plugin-version>3.1.0</maven-enforcer-plugin-version>
+    <build-helper-maven-plugin-version>3.3.0</build-helper-maven-plugin-version>
+    <maven-release-plugin-version>3.0.0-M7</maven-release-plugin-version>
+    <maven-source-plugin-version>3.2.1</maven-source-plugin-version>
+    <maven-javadoc-plugin-version>3.4.1</maven-javadoc-plugin-version>
+    <maven-deploy-plugin-version>3.0.0-M2</maven-deploy-plugin-version>
     <edge-api-utils-version>1.1.2</edge-api-utils-version>
-    <wiremock-version>2.33.2</wiremock-version>
+    <wiremock-version>2.35.0</wiremock-version>
     <!--Exclude this packages from sonar, as it contains only interfaces and entities-->
     <sonar.exclusions>
       src/main/java/org/folio/edgecommonspring/domain/**,
